myproject/requirements.txt
annotated-types
anyio
asgiref==3.7.2
cachetools==5.3.2
certifi
charset-normalizer==3.3.2
clovax==0.1.1
colorama
distro
Django==5.0.2
django-cors-headers
djangorestframework==3.14.0
exceptiongroup
google-ai-generativelanguage==0.4.0
google-api-core==2.17.0
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
h11
h2
hpack==4.0.0
httpcore
httpx
hyperframe
idna
openai
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic
pydantic_core
pytz==2024.1
requests==2.31.0
rest_framework_simplejwt
rsa==4.9
setuptools==69.0.3
sniffio
sqlparse==0.4.4
tqdm
typing_extensions
urllib3==2.2.0
wheel==0.42.0
